Moved MangleOpenCLBuiltinsPass to separate file
